requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
cohere==5.2.4
docx2txt==0.8
fake-useragent==1.5.1
greenlet==3.0.3
idna==3.6
joblib==1.3.2
langdetect==1.0.9
lxml==4.9.3
numpy==1.26.4
openai==1.30.1
pandas==2.2.2
pdfplumber==0.10.4
playwright==1.44.0
PyMuPDF==1.24.2
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.13.1
schedule==1.2.1
sentence-transformers==2.7.0
sqlalchemy==2.0.30
tqdm==4.66.2
urllib3==2.2.1
google-generativeai==0.5.4
